fix: create ui constants
