--- src/ui/__fixtures__/dictionary.ts
@@ -1,18 +1,0 @@
-const cardTypes = {
-  creds: "creds",
-  dids: "dids",
-};
-
-const connectionStatus = {
-  pending: "pending",
-  status: "confirmed",
-};
-
-const generateSeedPhraseState = {
-  onboarding: "onboarding",
-  additional: "additional",
-  restore: "restore",
-  success: "success",
-};
-
-export { cardTypes, connectionStatus, generateSeedPhraseState };
